login and registration routes working with their forms
