+++ templates/about.html
@@ -1,0 +1,3 @@
+{% extends "layout.html" %} {% block content %}
+<h1>About Page</h1>
+{% endblock content %}
