+++ templates/register.html
@@ -1,74 +1,92 @@
 <!--source: https://github.com/CoreyMSchafer/code_snippets/blob/master/Python/Flask_Blog/03-Forms-and-Validation/templates/register.html -->
+{% extends "layout.html" %} {% block content %}
 
-{% extends "layout.html" %}
-{% block content %}
-    <div class="content-section">
-        <form method="POST" action="">
-            {{ form.hidden_tag() }}
-            <fieldset class="form-group">
-                <legend class="border-bottom mb-4">Join Today</legend>
-                <div class="form-group">
-                    {{ form.username.label(class="form-control-label") }}
+<!-- Registration POSTS entry to Users table, in current location (action="")  -->
+<div class="content-section">
+	<form method="POST" action="">
+		<!-- Add CSRF token - form security -->
+		{{ form.hidden_tag() }}
 
-                    {% if form.username.errors %}
-                        {{ form.username(class="form-control form-control-lg is-invalid") }}
-                        <div class="invalid-feedback">
-                            {% for error in form.username.errors %}
-                                <span>{{ error }}</span>
-                            {% endfor %}
-                        </div>
-                    {% else %}
-                        {{ form.username(class="form-control form-control-lg") }}
-                    {% endif %}
-                </div>
-                <div class="form-group">
-                    {{ form.email.label(class="form-control-label") }}
-                    {% if form.email.errors %}
-                        {{ form.email(class="form-control form-control-lg is-invalid") }}
-                        <div class="invalid-feedback">
-                            {% for error in form.email.errors %}
-                                <span>{{ error }}</span>
-                            {% endfor %}
-                        </div>
-                    {% else %}
-                        {{ form.email(class="form-control form-control-lg") }}
-                    {% endif %}
-                </div>
-                <div class="form-group">
-                    {{ form.password.label(class="form-control-label") }}
-                    {% if form.password.errors %}
-                        {{ form.password(class="form-control form-control-lg is-invalid") }}
-                        <div class="invalid-feedback">
-                            {% for error in form.password.errors %}
-                                <span>{{ error }}</span>
-                            {% endfor %}
-                        </div>
-                    {% else %}
-                        {{ form.password(class="form-control form-control-lg") }}
-                    {% endif %}
-                </div>
-                <div class="form-group">
-                    {{ form.confirm_password.label(class="form-control-label") }}
-                    {% if form.confirm_password.errors %}
-                        {{ form.confirm_password(class="form-control form-control-lg is-invalid") }}
-                        <div class="invalid-feedback">
-                            {% for error in form.confirm_password.errors %}
-                                <span>{{ error }}</span>
-                            {% endfor %}
-                        </div>
-                    {% else %}
-                        {{ form.confirm_password(class="form-control form-control-lg") }}
-                    {% endif %}
-                </div>
-            </fieldset>
-            <div class="form-group">
-                {{ form.submit(class="btn btn-outline-info") }}
-            </div>
-        </form>
-    </div>
-    <div class="border-top pt-3">
-        <small class="text-muted">
-            Already Have An Account? <a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
-        </small>
-    </div>
+		<!-- form-group = Form's fieldset container  -->
+		<fieldset class="form-group">
+			<legend class="border-bottom mb-4">Join Today</legend>
+			<!-- Username  -->
+			<div class="form-group">
+                
+				{{ form.username.label(class="form-control-label") }} 
+                
+                {% if form.username.errors %}
+                    {{form.username(class="form-control form-control-lg is-invalid") }}
+				<div class="invalid-feedback">
+					{% for error in form.username.errors %}
+					<span>{{ error }}</span>
+					{% endfor %}
+				</div>
+				{% else %} 
+                    {{ form.username(class="form-control form-control-lg") }}
+				{%endif %}
+			</div>
+
+			<!-- Email -->
+			<div class="form-group">
+				{{ form.email.label(class="form-control-label") }} 
+
+                    {% if form.email.errors %} 
+                {{ form.email(class="form-control form-control-lg is-invalid") }}
+				<div class="invalid-feedback">
+					{% for error in form.email.errors %}
+					<span>{{ error }}</span>
+					{% endfor %}
+				</div>
+				{% else %} 
+                    {{ form.email(class="form-control form-control-lg") }}
+				{%endif %}
+			</div>
+
+			<!-- Password -->
+			<div class="form-group">
+				{{ form.password.label(class="form-control-label") }} 
+                
+                {% if form.password.errors %} 
+                {{ form.password(class="form-control form-control-lg is-invalid") }}
+				<div class="invalid-feedback">
+					{% for error in form.password.errors %}
+					<span>{{ error }}</span>
+					{% endfor %}
+				</div>
+				{% else %} 
+                    {{ form.password(class="form-control form-control-lg") }}
+				{%endif %}
+			</div>
+
+			<!-- Confirm Password -->
+			<div class="form-group">
+				{{ form.confirm_password.label(class="form-control-label") }} 
+                
+                {% if form.confirm_password.errors %} 
+                    {{ form.confirm_password(class="form-control form-control-lg is-invalid")}}
+				<div class="invalid-feedback">
+					{% for error in form.confirm_password.errors %}
+					<span>{{ error }}</span>
+					{% endfor %}
+				</div>
+				{% else %} 
+                    {{ form.confirm_password(class="form-control form-control-lg") }}
+                {% endif %}
+			</div>
+		</fieldset>
+
+		<!-- Submit button  -->
+		<div class="form-group">
+			{{ form.submit(class="btn btn-outline-info") }}
+		</div>
+	</form>
+</div>
+
+<div class="border-top pt-3">
+	<small class="text-muted">
+		Already Have An Account?
+		<a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
+	</small>
+</div>
 {% endblock content %}
